Fully converted all code over the database.

+++ pad_db/format.rs
@@ -15,7 +15,7 @@ pub(super) struct Axis {
 pub(super) struct Trigger {
     pub(super) code: u8,
     pub(super) event: String,
-    pub(super) max: Option<u32>,
+    pub(super) max: Option<f64>,
     pub(super) deadzone: Option<f64>,
 }
